// app/lib/types.ts
export type AppMode = "emoji" | "paint" | "pixel" | "pixelate" | "formula";
export type ColorMode = "single" | "linear" | "radial";
export type PlacementMode = "random" | "grid";
export type GradientDirection =
  | "diagonal-tl-br"
  | "diagonal-tr-bl"
  | "diagonal-bl-tr"
  | "diagonal-br-tl"
  | "horizontal-lr"
  | "horizontal-rl"
  | "vertical-tb"
  | "vertical-bt";
export type ImageFormat = "jpeg" | "png";
export type PixelateFit = "cover" | "contain";